pyproject.toml
[project]
name = "mspr2-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "whitenoise>=6.11.0",
]
